server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-cors = "^5.0.0"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"
flask-socketio = "^5.3.7"
eventlet = "^0.36.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
